+++ influxdb_client/client/users_api.py
@@ -5,7 +5,8 @@ To grant a user permission to access data, add them as a member of an organizati
 and provide them with an authentication token.
 """
 
-from influxdb_client import UsersService, User
+from typing import Union
+from influxdb_client import UsersService, User, Users, UserResponse
 
 
 class UsersApi(object):
@@ -26,3 +27,39 @@ class UsersApi(object):
         user = User(name=name)
 
         return self._service.post_users(user=user)
+
+    def update_user(self, user: User) -> UserResponse:
+        """Update a user.
+
+        :param user: User update to apply (required)
+        :return: User
+        """
+        return self._service.patch_users_id(user_id=user.id, user=user)
+
+    def delete_user(self, user: Union[str, User, UserResponse]) -> None:
+        """Delete a user.
+
+        :param user: user id or User
+        :return: User
+        """
+        if isinstance(user, User):
+            user_id = user.id
+        elif isinstance(user, UserResponse):
+            user_id = user.id
+        else:
+            user_id = user
+
+        return self._service.delete_users_id(user_id=user_id)
+
+    def find_users(self, **kwargs) -> Users:
+        """List all users.
+
+        :key int offset: Offset for pagination
+        :key int limit: Limit for pagination
+        :key str after: The last resource ID from which to seek from (but not including).
+                        This is to be used instead of `offset`.
+        :key str name: The user name.
+        :key str id: The user ID.
+        :return: Buckets
+        """
+        return self._service.get_users(**kwargs)