+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 1.24.0 [unreleased]
 
+### Features
+1. [#358](https://github.com/influxdata/influxdb-client-python/pull/358): Update management API:
+   - `BucketsApi` - add possibility to: `update`
+   - `OrganizationsApi` - add possibility to: `update`
+   - `UsersApi` - add possibility to: `update`, `delete`, `find`
+
 ## 1.23.0 [2021-10-26]
 
 ### Deprecates