+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.24.0 [unreleased]
+## 1.24.0 [2021-11-26]
 
 ### Features
 1. [#358](https://github.com/influxdata/influxdb-client-python/pull/358): Update management API: